server/requirements.txt
# FastAPI and Server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# PDF Processing
pypdf>=3.0.0
reportlab>=4.0.0
pdf2image>=1.16.0

# Image Processing
Pillow>=10.0.0

# Document Processing
python-docx>=1.0.0
python-pptx>=0.6.0
openpyxl>=3.1.0

# File Handling
aiofiles>=23.0.0

# Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
